HYHULL-5: Support Continuous Integration Testing with Hudson
Switched from hydra-testing to jettywrapper gem.

Swapped pattern of 
	Rake::Task["foo:bar"].invoke if index == 0
    Rake::Task["foo:bar"].execute if index > 0
to
	Rake::Task["foo:bar"].reenable
	Rake::Task["foo:bar"].invoke
as the former would cause the first invoke task to
be skipped when tasks were chained together as they
are in hull:hudson.

As an expediency, created the task hull:jettycleaner
but this should be replaced when we unify fixture 
handling in a gem.

https://jira.duraspace.org/browse/HYHULL-5